consertei a busca de cidade por estado, usar Long e nao int e acessar o atributo e nao o objeto na consulta

+++ src/main/java/br/edu/pos/confrov/dao/ICidadeDAO.java
@@ -12,7 +12,7 @@ public interface ICidadeDAO {
 
 	public Cidade findByNome(String nome);
 	
-	public List<Cidade> findbyEstado (int estado);
+	public List<Cidade> findbyEstado (Long estado);
 	
 	public List<Cidade> findByAll();
